backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "sse-starlette[daphne,examples,granian,uvicorn]>=3.1.2",
    "uvicorn>=0.40.0",
]
